requirements.txt
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
Pillow==7.2.0
opencv-python==4.4.0.44
python-xlib==0.28
imutils==0.5.3
requests==2.24.0
motor== 2.3.0
pymongo==3.11.0
